consistent config structure in autopaho
The autopaho config was a mix of public fields
and private fields with setters. This change makes
everything public in line with paho. The setters
have been left for compatability.

Also remove the term "broker" because this is
not used in the MQTT spec so is likely to be
confusing.

Closes #174